move code to fix circular dependency graph
