+++ packages/zero-schema/package.json
@@ -14,6 +14,7 @@
     "lint": "eslint --ext .ts,.tsx,.js,.jsx src/"
   },
   "devDependencies": {
+    "zero-protocol": "0.0.0",
     "zql": "0.0.0"
   }
 }
