+++ packages/zero-schema/src/table-schema.ts
@@ -1,6 +1,5 @@
+import type {ValueType} from '../../zero-protocol/src/client-schema.ts';
 import type {PrimaryKey} from '../../zero-protocol/src/primary-key.ts';
-
-export type ValueType = 'string' | 'number' | 'boolean' | 'null' | 'json';
 
 /**
  * `related` calls need to know what the available relationships are.
